frontend, eval gate (all green), unit tests, docs, release tooling
- three-view SPA (Studio timeline + DNA barcode / Atlas portals + walk /
  Bodies + URDF dropzone), dark + papaya, zero build step
- release gate self-contained on packaged artifacts, ALL GREEN:
  reencode 0.64, follow 0.42, atlas 6.1x, bridge excess 1.27 (bar 1.5,
  measured extrapolation cost documented), retime 0.87, synergy 61%
- 10 unit tests incl. forbidden-path portability blocker
- public README (differentiation: native cross-embodiment, Atlas Map,
  constraint-native editor, generative library, self-grading) + SDK doc
- HF model card + upload script (blocked on a write token: current token
  is read-only)

+++ src/alphamotion/engine/equator.py
@@ -112,7 +112,7 @@ class Equator:
     def token_nll(self, tokens: torch.Tensor, endpoints, n_frames: int):
         """Per-token NLL of a GIVEN sequence under B3 (full mask) — the
         synergy-gate ruler."""
-        tok = tokens[None].to(self.device)
+        tok = tokens[None].long().to(self.device)
         mask = torch.ones_like(tok, dtype=torch.bool)
         nf = torch.tensor([float(n_frames)], device=self.device)
         logits = self.b3(tok, mask, endpoints, nf)
